test: test coverage for utils.attributes
This initial commit is primarily to show a strategy for converting
abstract class attributes accessed by method calls to use simple
attribute access instead. It turned out to be a little more involved
than I'd hoped, mostly due to the requirement of preserving backwards
compatibility. ("We do these things, not because they are easy, but
because we thought they would be easy.")

This commit uses the ``name`` attribute as the first example for a
working strategy going forward.

The primary innovation here is the introduction of the
``abstract_class_property`` class that is used to define abstract class
properties. This is implemented independently of the ``abc`` package in
the standard library, since it has no support for abstract class
attributes.  One thing that's a little weird is figuring out when to
check for concrete values for abstract attributes. With instance methods
and attributes, the ``abc`` package just waits until instantiation to
check the entire class to make sure it's concrete before instantiating.
We do that, too, for these class properties, but since a class property
can be accessed independtly of an instance, we also check individual
properties as they are accessed. Some mechanisms for checking subclasses
as they are declared were tried and discarded, as there's no real good
way to know when the user has finished subclassing an abstract base
class, unless they try to instantiate it.

As far as backwards compatibility, the first difficulty is we can't
simply convert calls to ``DatasetManager.name()`` to
``DatasetManager.name``, since that would break downstream code that
still uses the method call syntax. Unfortunately, that means that for
each attribute we want to convert we need to introduce a new name. In
this case ``name`` has been migrated to ``dataset_name``. Code which
uses the ``name`` class method will generate a DeprecationWarning.
Generally speaking, we should try to get rid of those warnings as they
crop up in downstream code, by migrating that code sooner than later.
When defining the ``dataset_name`` abstract class property, ``"name"``
is passed for the ``fallback`` argument. When the ``dataset_name``
property is accessed, first it will look for a concrete definition for
``dataset_name`` in the bound class. If it fails to find that, then it
will fall back to looking for a ``name`` class method and use that (with
a DeprecationWarning) if it finds one.

There is also a, seemingly circular, definition of the ``name``
classmethod in the ``Attributes`` mixin which uses the
``_find_fallback`` method to look for a ``dataset_name`` value. This is
not for backwards compatibility with downstream code, as dataset
managers that use ``name`` will also have defined ``name``, so this
won't be used for those. This is here because ``gridded_etl_tools``
doesn't yet have 100% test coverage, so there is always some risk of
some untested code in ``gridded_etl_tools`` using an older style class
method instead of the property which has replaced it. When test coverage
is up to 100%, it should be possible to remove ``_find_fallback`` and
the fallback class methods that use it.