Cover line 11.
Possibly we wouldn't need the "pragma NO COVER" if we incorporated type
checking into the unit testing, but that is uncharted territory for me
and not necessary at this juncture. For now, we can just acknowledge
that we know it's not covered, which will make the coverage tool not
count line 11 against us.