Update filtering and scraping in HTTP extractor
The new filter_func argument to the link scraping function accepts a
generic function that receives the link href string and returns a
boolean indicating whether or not to filter the link out.

The code for link scraping is moved into HTTPExtractor and renamed from
get_hrefs to get_links because it now returns the absolute URL of each
link instead of literally its href value.

HTTPExtractor creates a requests.session object, freeing up the
caller from having to manage one of its own. HTTPExtractor is now
required to be created within a context manager, which it uses to
automatically open and close the session.

In an effort to make HTTPExtractor.request work smoothly with
Extractor.pool when requesting a list of urls to download,
Extractor.pool now accepts 1D arguments in addition to 2D ones. 1D
arguments are passed one by one to Extractor.pool.

Corresponding tests for HTTPExtractor and Extractor.pool are added to
test_extractor. A new testing dependency is added: the "responses"
Python package. It mocks responses from HTTP servers, making it
possible to test requests for link scraping, filtering, and
downloading.