Remove dm.local_input_path from HTTPExtractor
This changes HTTPExtractor.request so files are written to the working
directory by default instead of to DatasetManager.local_input_path. This
allows HTTPExtractor to be less coupled with DatasetManager, and it
gives it similar behavior to FTPExtractor.

The mock HTTP requests are modified to change into a temp directory to
test the new behavior.

Another small change is the HTTP content is extracted before the
destination file is created. This prevents an empty file from being
left behind when the HTTP request fails.