Improve automatic HTTP request retries
This changes the default parameters for HTTP request retries to more
reasonable defaults of a maximum 2 minute wait between the last two
requests. Previously, if all the requests failed, the final request
would wait for over an hour before running.

The tests have been updated to test that the correct number of
retries are running for requests that don't succeed at first.